Trying to increase float sizes

+++ trainTree.py
@@ -14,7 +14,7 @@ def read_glove(file_name):
     for line in file(file_name):
         line = line.split()
         word = line.pop(0)
-        W2NV[word] = np.array(map(float, line)).astype(np.float128)
+        W2NV[word] = np.array(map(np.float128, line))
     return W2NV
